Load tests using unified argument spec
This style follows what is normal convention in nosetest and other
testing frameworks. The syntax is package:class.function. In our case,
since we allow testing without a class, we also allow package:class or
package:function.

Also improved was the output of the actual test name, which allows it to
be easily repeatable in the event of an error

The --collect-only option allows to print the tests detected, without
actually running any of them.