Change loader to return serializable objects (#150)
* Change loader to return serializable objects

Now that it is possible for tests to run in parallel
it require serialize and deserialize the test to another
process. For this to work the test must be serializable.

Most of the time it works fine but sometimes (for example
when tests are used with decorator) it will not work.

Change loader to return a serializable test. The returned
object will hold a metadata about the test (file, and symbol).
Each processes will use this information to load the test on
spot and run it.

* review fixes

* allow ask for a fresh env even when reuse-env is used